Move STO logic into CMPTransaction file
